Automatically enable WiFi if it's determined safe.
When running this script on a fresh OpenWrt setup, it should be safe
to automatically bring the new SSID(s) online. The same is true for
setups that don't have any other enabled SSIDs configured anyway.

Just in case another SSID is found that might be offline on purpose,
refrain from enabling the affected radio (if it's disabled, that is).

There's no behavior change if your radios were already enabled in the
first place - they will remain enabled with the new SSID(s) starting
to broadcast.